refactor: Extract PublishDraftButton and PublishedChip components
