+++ src/library-authoring/containers/index.scss
@@ -1,2 +1,1 @@
 @import "./ContainerCard.scss";
-@import "./ContainerPublishStatus.scss";
