+++ src/library-authoring/generic/index.scss
@@ -1,3 +1,4 @@
 @import "./history-widget/HistoryWidget";
 @import "./status-widget/StatusWidget";
 @import "./parent-breadcrumbs";
+@import "./publish-status-buttons";
