+++ src/library-authoring/generic/publish-status-buttons/index.scss
@@ -1,20 +1,3 @@
-.status-box {
-  border: 2px solid;
-  border-radius: 4px;
-
-  &.draft-status {
-    @extend %draft-status;
-  }
-
-  &.published-status {
-    @extend %published-status;
-  }
-
-  .container-name {
-    width: 200px;
-  }
-}
-
 .status-button {
   border: 1px solid;
   border-left: 4px solid;
