+++ src/library-authoring/generic/publish-status-buttons/index.tsx
@@ -1,0 +1,2 @@
+export { PublishedChip } from './PublishedChip';
+export { PublishDraftButton } from './PublishDraftButton';
