+++ src/library-authoring/index.scss
@@ -5,7 +5,7 @@
 @import "./units";
 @import "./section-subsections";
 @import "./containers";
-@import "./hierarchy/ItemHierarchy";
+@import "./hierarchy";
 
 .library-cards-grid {
   display: grid;
